Refactor: move challenge localization to challenge-transformer
